finalize backend

+++ components/Post.tsx
@@ -8,7 +8,7 @@ import { formatDistanceToNow } from 'date-fns'
 import { Image } from 'expo-image'
 import { Link } from 'expo-router'
 import React, { useState } from 'react'
-import { Text, TouchableOpacity, View } from 'react-native'
+import { Alert, Text, TouchableOpacity, View } from 'react-native'
 import { styles } from '../styles/feed.styles'
 import CommentsModal from './CommentsModal'
 
@@ -34,8 +34,6 @@ export default function Post({ post }: PostProps ) {
 
     const [isLiked, setIsLiked] = useState(post.isLiked)
     const [isBookmarked, setIsBookmarked] = useState(post.isBookmarked)
-    const [likesCount , setLikesCount] = useState(post.likes)
-    const [commentsCount , setCommentsCount] = useState(post.comments)
     const [showComments, setShowComments] = useState(false)
 
     const {user} = useUser() //clerk user doesn't have spesific fields like in the db
@@ -45,19 +43,36 @@ export default function Post({ post }: PostProps ) {
     const toggleBookmark = useMutation(api.bookmarks.toggleBookmark)
     const deletePost = useMutation(api.posts.deletePosts)
 
-    const handleDelete = async () => {
-        try {
-            await deletePost({postId: post._id})
-        }catch (error) {
-            console.log( "Error deleting post:" , error)
+    const handleDelete = () => {
+    Alert.alert(
+        "Delete Post",
+        "Are you sure you want to delete this post?",
+        [
+        {
+            text: "Cancel",
+            style: "cancel"
+        },
+        {
+            text: "Confirm",
+            onPress: async () => {
+            try {
+                await deletePost({ postId: post._id });
+                console.log("Post deleted successfully");
+            } catch (error) {
+                console.log("Error deleting post:", error);
+            }
+            },
+            style: "destructive"
         }
-    }
+        ],
+        { cancelable: true }
+    );
+    };
 
     const handleLike = async () => {
         try {
             const newIsLiked =await toggleLike({ postId: post._id })
             setIsLiked(newIsLiked)
-            setLikesCount((prev) => newIsLiked ? prev + 1 : prev - 1)
         }catch (error) {
             console.log( "Error liking post:" , error)
         }
@@ -70,7 +85,9 @@ export default function Post({ post }: PostProps ) {
     return (
         <View style={styles.post}>
             <View style={styles.postHeader}>
-                <Link href={"/(tabs)"}>
+                <Link href={
+                    post.author._id === currentUser?._id ? '/(tabs)/profile' : `/user/${post.author._id}`
+                } asChild>
                     <TouchableOpacity style={styles.postHeaderLeft}>
                         <Image source={post.author.image} style={styles.postAvatar} contentFit="cover" transition={200}
                         cachePolicy="memory-disk" />
@@ -112,16 +129,16 @@ export default function Post({ post }: PostProps ) {
             </View>
             {/* post info */}
             <View style={styles.postInfo}>
-                <Text style={styles.likesText}>{likesCount > 0 ? `${likesCount.toLocaleString()} likes`: "Be the first to like" }</Text>
+                <Text style={styles.likesText}>{post.likes > 0 ? `${post.likes.toLocaleString()} likes`: "Be the first to like" }</Text>
                 {post.caption && (
                     <View style={styles.captionContainer}>
                         <Text style={styles.captionUsername}>{post.author.userName}</Text>
                         <Text style={styles.captionText}>{post.caption}</Text>
                     </View>
                 )}
-                {commentsCount > 0 && (
+                {post.comments > 0 && (
                     <TouchableOpacity onPress={() => setShowComments(true)}>
-                        <Text style={styles.commentsText}>View all {commentsCount} comments</Text>
+                        <Text style={styles.commentsText}>View all {post.comments} comments</Text>
                     </TouchableOpacity>
                 )}
                 <Text style={styles.timeAgo}>
@@ -133,8 +150,6 @@ export default function Post({ post }: PostProps ) {
                 postId = {post._id}
                 visible={showComments}
                 onClose={() => setShowComments(false)}
-                onCommentAdded={() => setCommentsCount(prev => prev + 1)}
-
             />
         </View>
     )
